% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictorSel.R
\name{modelSel_rf}
\alias{modelSel_rf}
\title{select predictors in randomForest}
\usage{
modelSel_rf(f, train_data, valid_x, valid_y, acc_thresh = 0.99)
}
\arguments{
\item{f}{formula}

\item{train_data}{the dataframe for training data set}

\item{valid_x}{the dataframe for predictors in valid data set}

\item{valid_y}{the vector for responser for valid data set}

\item{acc_thresh}{the numeric value of the accuracy thresh hold to stop}
}
\value{
a list of the predictors' names
}
\description{
select predictors in randomForest
}
\details{
This function return the name list for predictors in RandomForest with the most importance calculated by varImp()
and accuracy higher than acc_thresh.
}
\author{
wangh
}
\seealso{
\code{randomForest}, \code{confusionMatrix}, \code{varImp}
}

